requirements/base.txt
nltk==3.3
gitpython==2.1.11